+++ server/libs/passport/strategies/jwt.js
